// Shelf.WebApp/ShelfApp/src/app/shared/models/store/stores.ts
export class Stores {
    Table: Store[];
}
export class Store {
    DivisionName: string;
    StoreNum: number;
    IDStore: number;
    LKInternalStatus: number;
    EffectiveFrom: any;
    EffectiveTo: any;
    Name: string
    Address1: string;
    Address2: string;
    City: string;
    State: string;
    PostalCode: string;
    Country: string;
    Phone: string;
    Email: string;
    Fax: string;
    L1: string;
    L2: string;
    L3: string;
    L4: string;
    L5: string;
    L6: string;
    L7: string;
    L8: string;
    L9: string;
    L10: string;
    StoreSize: string;
    RemodelDate: any;
    MerchareaSize: number
    DistrictCode: string
    FacilityName: string
    ROG: any
    IDNotes: number;
    TotalArea: number;
    MerchArea: number;
    IsActive: boolean;
    CreatedTS: any;
    CreatedBy: string;
    ModifiedTS: any;
    ModifiedBy: string;
    Latitude: number;
    Longitude: number;
    PogCount: number;
    StoreDescX01: string;
    StoreDescX02: string;
    StoreDescX03: string;
    StoreDescX04: string;
    StoreDescX05: string;
    selected?:boolean;
}